changed monster classes names
