+++ src/main/java/chronometry/monsters/AcidSlimeMPatch.java
@@ -4,27 +4,27 @@ import basemod.ReflectionHacks;
 import chronometry.patches.AbstractMonsterPatch;
 import chronometry.IntentData;
 import chronometry.effects.AttackEffect;
-import chronometry.effects.DebuffFrailEffect;
+import chronometry.effects.DebuffWeakEffect;
 import chronometry.effects.DebuffCardDiscardEffect;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.cards.status.Slimed;
 import com.megacrit.cardcrawl.monsters.AbstractMonster.Intent;
-import com.megacrit.cardcrawl.monsters.exordium.SpikeSlime_M;
+import com.megacrit.cardcrawl.monsters.exordium.AcidSlime_M;
 
 import java.util.ArrayList;
 
-public class SpikeSlimeMMoves {
-    @SpirePatch(clz=SpikeSlime_M.class,
+public class AcidSlimeMPatch {
+    @SpirePatch(clz=AcidSlime_M.class,
                 method=SpirePatch.CONSTRUCTOR,
-                paramtypez = {float.class, float.class})
-    public static class SpikeSlimeMInitMoves {
-        public static void Postfix(SpikeSlime_M __instance, float x, float y) {
+                paramtypez = {float.class, float.class, int.class, int.class})
+    public static class InitMoves {
+        public static void Postfix(AcidSlime_M __instance, float x, float y, int poisonAmount, int newHealth) {
             ArrayList<IntentData> moves = new ArrayList<IntentData>();
 
             IntentData move1 = new IntentData(
-                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "FLAME_TACKLE"),
+                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "WOUND_TACKLE"),
                     Intent.ATTACK_DEBUFF,
-                    "Flame Tackle"
+                    AcidSlime_M.MOVES[0]
             );
             move1.add_effect(new AttackEffect(
                     __instance.damage.get(0),
@@ -36,12 +36,23 @@ public class SpikeSlimeMMoves {
             ));
             moves.add(move1);
 
-            IntentData move4 = new IntentData(
-                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "FRAIL_LICK"),
-                    Intent.DEBUFF,
-                    SpikeSlime_M.MOVES[0]
+            IntentData move2 = new IntentData(
+                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "NORMAL_TACKLE"),
+                    Intent.ATTACK,
+                    "Normal Tackle"
             );
-            move4.add_effect(new DebuffFrailEffect(
+            move2.add_effect(new AttackEffect(
+                    __instance.damage.get(1),
+                    1
+            ));
+            moves.add(move2);
+
+            IntentData move4 = new IntentData(
+                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "WEAK_LICK"),
+                    Intent.DEBUFF,
+                    AcidSlime_M.MOVES[1]
+            );
+            move4.add_effect(new DebuffWeakEffect(
                     1
             ));
             moves.add(move4);
