+++ src/main/java/chronometry/monsters/SpikeSlimeMPatch.java
@@ -4,27 +4,27 @@ import basemod.ReflectionHacks;
 import chronometry.patches.AbstractMonsterPatch;
 import chronometry.IntentData;
 import chronometry.effects.AttackEffect;
-import chronometry.effects.DebuffWeakEffect;
+import chronometry.effects.DebuffFrailEffect;
 import chronometry.effects.DebuffCardDiscardEffect;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.cards.status.Slimed;
 import com.megacrit.cardcrawl.monsters.AbstractMonster.Intent;
-import com.megacrit.cardcrawl.monsters.exordium.AcidSlime_M;
+import com.megacrit.cardcrawl.monsters.exordium.SpikeSlime_M;
 
 import java.util.ArrayList;
 
-public class AcidSlimeMMoves {
-    @SpirePatch(clz=AcidSlime_M.class,
+public class SpikeSlimeMPatch {
+    @SpirePatch(clz=SpikeSlime_M.class,
                 method=SpirePatch.CONSTRUCTOR,
-                paramtypez = {float.class, float.class, int.class, int.class})
-    public static class AcidSlime_MInitMoves {
-        public static void Postfix(AcidSlime_M __instance, float x, float y, int poisonAmount, int newHealth) {
+                paramtypez = {float.class, float.class})
+    public static class InitMoves {
+        public static void Postfix(SpikeSlime_M __instance, float x, float y) {
             ArrayList<IntentData> moves = new ArrayList<IntentData>();
 
             IntentData move1 = new IntentData(
-                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "WOUND_TACKLE"),
+                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "FLAME_TACKLE"),
                     Intent.ATTACK_DEBUFF,
-                    AcidSlime_M.MOVES[0]
+                    "Flame Tackle"
             );
             move1.add_effect(new AttackEffect(
                     __instance.damage.get(0),
@@ -36,23 +36,12 @@ public class AcidSlimeMMoves {
             ));
             moves.add(move1);
 
-            IntentData move2 = new IntentData(
-                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "NORMAL_TACKLE"),
-                    Intent.ATTACK,
-                    "Normal Tackle"
-            );
-            move2.add_effect(new AttackEffect(
-                    __instance.damage.get(1),
-                    1
-            ));
-            moves.add(move2);
-
             IntentData move4 = new IntentData(
-                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "WEAK_LICK"),
+                    (byte)ReflectionHacks.getPrivateStatic(__instance.getClass(), "FRAIL_LICK"),
                     Intent.DEBUFF,
-                    AcidSlime_M.MOVES[1]
+                    SpikeSlime_M.MOVES[0]
             );
-            move4.add_effect(new DebuffWeakEffect(
+            move4.add_effect(new DebuffFrailEffect(
                     1
             ));
             moves.add(move4);
